Moving Composer helpers to their own files
