+++ EssentialFeediOS/Feed UI/Composers/FeedLoaderPresentationAdapter.swift
@@ -1,0 +1,32 @@
+//
+//  FeedLoaderPresentationAdapter.swift
+//  EssentialFeediOS
+//
+//  Created by Macbook Pro on 21/9/23.
+//
+
+import Foundation
+import EssentialFeed
+
+private final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
+    private let feedLoader: FeedLoader
+    var presenter: FeedPresenter?
+    
+    init(feedLoader: FeedLoader) {
+        self.feedLoader = feedLoader
+    }
+    
+    func didRequestFeedRefresh() {
+        presenter?.didStartLoadingFeed()
+        
+        feedLoader.load { [weak self] result in
+            switch result {
+                case let .success(feed):
+                    self?.presenter?.didFinishLoadingFeed(with: feed)
+                    
+                case let .failure(error):
+                    self?.presenter?.didFinishLoadingFeed(with: error)
+            }
+        }
+    }
+}
